% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdatemplate.R
\docType{package}
\name{fdatemplate}
\alias{fdatemplate}
\alias{fdatemplate-package}
\title{package to distribute fdatemplate}
\description{
Open a new rmarkdown file.  Select the "fdatemplate" from the list of templates.
}
